Backend/requirements.txt
# Create requirements.txt content
@"
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
requests==2.31.0
python-multipart==0.0.6
Pillow==10.1.0
torch==2.1.1
transformers==4.35.2
langchain==0.0.354
pinecone-client==3.0.1
sentence-transformers==2.2.2
python-dotenv==1.0.0
numpy==1.24.3
"@ | Out-File -FilePath "requirements.txt" -Encoding UTF8

streamlit
pillow
requests
fastapi
uvicorn
numpy
pydantic
python-multipart

fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
openpyxl==3.1.2
python-multipart==0.0.6
